Make slice a mutation free operation.
Due to the way MobX works creating a new MArray by making an empty
array then mutating it to get to a result will cause a failure in
its cycle detection heuristics.

See a simple example here: https://github.com/mobxjs/mobx/issues/1684

To make sure this issue isn't a problem, we make all methods that
don't mutate the source MArray use the MArray constructor. The constructor
is safe as it doesn't subsequently mutate the newly created observable
values, nor it tries to read them.

This PR applies the fix to the slice method; previous PRs applied it
to map, filter and similar methods. A realworld test was added to
simulate this scenario which should ensure we dont have a regression